refactor: replace array helper functions with new implementations in helpers package
